requirements.txt
# Methane Plume Detection System - Requirements
# Optimized for AMD 5600U (CPU-based training)

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and CLI
tqdm>=4.65.0

# Energy and Carbon Tracking
codecarbon>=2.3.0

# Demo Interface
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0

# Scientific Computing
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: Experiment Tracking
# wandb>=0.15.0

# Optional: Jupyter Notebooks
# jupyter>=1.0.0
# ipykernel>=6.25.0

# Development Tools (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0
